requirements.txt
ollama
langchain
langchain-core
langchain-community
langchain-ollama
jupyter
gtts
openai-whisper
SpeechRecognition


